Move findbyids to server-side-only
